#5 refactor: NavBar Item들을 직접 넘겨주는 방식으로 수정
